requirements.txt
aerich==0.6.3
aiofiles==0.8.0
aiomysql==0.1.1
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.79.0
fastapi-mail==1.0.9
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iso8601==1.0.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
protobuf==4.21.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
redis==4.3.4
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
tomlkit==0.11.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uvicorn==0.18.2
watchfiles==0.16.0
websockets==10.3
wrapt==1.14.1